![breaking]: eliminate TinaContext
After about 3 or 4 itereation on the design intent for TinaContext, the
current design/implementation means its no longer useful and I can
eliminate it. WHile it's still a stackless Isolate handler, the new
Isolate_Turn_Frame stores a few framework related data needed for the
turn/quantum during execution. This feels/reads better. Maybe the next
revision I can further make it simpler. However, this is better and
simpler!

+++ src/simulated_test_determinism.odin
@@ -92,7 +92,7 @@ when TINA_SIMULATION_MODE {
 		simulator_run(&sim)
 
 		shard := &sim.shards[0]
-		ping_pointer := _get_isolate_ptr(shard, u16(PING_TYPE_ID), 0)
+		ping_pointer := _get_isolate_ptr(shard, PING_TYPE_ID, 0)
 		ping_memory := cast(^PingIsolate)ping_pointer
 
 		return Ping_Pong_Run_Result {
@@ -212,7 +212,7 @@ when TINA_SIMULATION_MODE {
 			simulator_run(&sim)
 
 			shard := &sim.shards[0]
-			ping_pointer := _get_isolate_ptr(shard, u16(PING_TYPE_ID), 0)
+			ping_pointer := _get_isolate_ptr(shard, PING_TYPE_ID, 0)
 			ping_memory := cast(^PingIsolate)ping_pointer
 
 			return Ping_Pong_Run_Result {
